ansible-playbook を実行するようにした

+++ scripts/install-extras.sh
@@ -1,2 +1,14 @@
 #!/bin/bash
-sudo apt-get install ansible -y 
+set -x
+
+if [[ ! -f /.install-extras ]]; then
+  sudo apt-get install ansible -y
+
+  (
+    export ANSIBLE_LOCALHOST_WARNING=False
+    export ANSIBLE_INVENTORY_UNPARSED_WARNING=False
+    ansible-playbook --diff /vagrant/ansible/playbook.yml
+  )
+
+  touch /.install-extras
+fi
